% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_edge_swap.R
\name{bipartite_edge_swap}
\alias{bipartite_edge_swap}
\title{Swap edges of a bipartite graph}
\usage{
bipartite_edge_swap(gr, Q, N = Q * igraph::ecount(gr))
}
\arguments{
\item{gr}{a tidygraph object with a node attribute called \code{type} that
holds boolean values (i.e. either \code{TRUE} or \code{FALSE})}

\item{Q}{number of permutations you are conducting}

\item{N}{the number of edge swaps; default is \eqn{Q \times |E(G)|}}
}
\value{
the graph woth \eqn{N} random edge swaps with constrained marginals
}
\description{
Run the edge swapping algorithm from Milo et al. (2003) on a
  bipartite graph while maintaing the separation of the nodes intwo groups.
}
\examples{
set.seed(0)
bgr <- tidygraph::create_ring(6, directed = FALSE)
bgr <- tidygraph::mutate(bgr, type = rep(c(TRUE, FALSE), 3))
print(bgr)
print(bipartite_edge_swap(bgr, 10))

}
