% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_edge_swap2.R
\name{bipartite_edge_swap2}
\alias{bipartite_edge_swap2}
\title{Swap edges of a bipartite graph (method 2)}
\usage{
bipartite_edge_swap2(gr, Q = 20, N = igraph::ecount(gr) * Q)
}
\arguments{
\item{gr}{a tidygraph object with a node attribute called \code{type} that
holds boolean values (i.e. either \code{TRUE} or \code{FALSE})}

\item{Q}{number of permutations you are conducting; default is 20}

\item{N}{the number of edge swaps; default is \eqn{Q \times |E(G)|}}
}
\value{
the graph with \eqn{N} random edge swaps with constrained marginals
}
\description{
Run the edge swapping algorithm from Milo et al. (2003) on a
  bipartite graph while maintaining the separation of the nodes in two groups.
}
\examples{
set.seed(0)
bgr <- tidygraph::create_ring(6, directed = FALSE)
bgr <- tidygraph::mutate(
    bgr,
    type = rep(c(TRUE, FALSE), 3),
    name = LETTERS[1:6])
print(bgr)
print(bipartite_edge_swap2(bgr, 10))

}
