% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_events.R
\name{calculate_events}
\alias{calculate_events}
\alias{calc_mutex_events}
\alias{calc_comut_events}
\title{Calculate the number of mutually exclusive or co-mutation events}
\usage{
calc_mutex_events(gs, bgr, sample_is = TRUE)

calc_comut_events(gs, bgr, sample_is = TRUE)
}
\arguments{
\item{gs}{the "gene set" is a vector of genes to consider}

\item{bgr}{a bipartite graph (as a tidygraph graph object); it must have a
boolean node attribute "type" that describes which group each node is in}

\item{sample_is}{either \code{TRUE} or \code{FALSE} dictating which boolean
value corresponds to the samples in the "type" attribute of \code{bgr}}
}
\description{
Calculate the number of mutually exclusive or co-mutation events
  in a bipartite graph
}
\examples{
set.seed(0)
suppressPackageStartupMessages(library(tidygraph))
bgr <- make_sample_gene_bipartite(
    simple_dataset$sample_name,
    simple_dataset$mutated_gene
)
plot(bgr)
calc_mutex_events(gs = c("a", "b"), bgr)
calc_comut_events(gs = c("c", "d"), bgr)

}
