% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_row_col_exclusivity_weights.R
\name{calculate_row_col_exclusivity_weights}
\alias{calculate_row_col_exclusivity_weights}
\title{Calculate the weights for the Row-Column-Exclusivity test}
\usage{
calculate_row_col_exclusivity_weights(dat, sample_col, mutgene_col,
  Q = 100)
}
\arguments{
\item{dat}{tibble with mutation information}

\item{sample_col}{column of samples names (quoted)}

\item{mutgene_col}{column of genes that are mutated (quoted)}

\item{Q}{number of permutation matrices to use (default is 100)}
}
\description{
The weight for each mutation in the Row-Column-Exclusivity test
  cannot be calculated exactly for most use cases because there is no closed
  formula and a very large number of possible matrices to consider. Thus, the
  weights are calculated emperically from a sufficiently large number of
  samples of possible matrices.
}
\examples{
library(wext)
calculate_row_col_exclusivity_weights(dat = simple_dataset,
                                      sample_col = sample_name,
                                      mutgene_col = mutated_gene,
                                      Q = 5)

}
