% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_row_exclusivity_weights.R
\name{calculate_row_exclusivity_weights}
\alias{calculate_row_exclusivity_weights}
\title{Calculate the weights for the Row-Exclusivity test}
\usage{
calculate_row_exclusivity_weights(dat, sample_col, mutgene_col)
}
\arguments{
\item{dat}{tibble with mutation information}

\item{sample_col}{column of samples names (quoted)}

\item{mutgene_col}{column of genes that are mutated (quoted)}
}
\description{
The weight for each gene in the Row-Exclusivity test is the
  number of samples with a mutation in a gene divided by the number of
  samples.
}
\examples{
library(wext)
calculate_row_exclusivity_weights(dat = simple_dataset,
                                  sample_col = sample_name,
                                  mutgene_col = mutated_gene)

}
