% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_gr}
\alias{check_gr}
\title{Check that a node attribute exists for a graph}
\usage{
check_gr(gr, node_attr)
}
\arguments{
\item{gr}{a tidygraph graph object}

\item{node_attr}{name of node attribute to check for}
}
\value{
(silently) TRUE or throws an error with the message: "<node_attr> not
  a node attribute in the graph"
}
\description{
If the node attribute exists, \code{TRUE} is returned
  silently; otherwise the function throws an informative error.
}
\examples{
gr <- tidygraph::create_ring(5)
gr <- tidygraph::mutate(gr, new_attr = LETTERS[1:5])
check_result <- check_gr(gr, "new_attr")
print(check_result)

}
