% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_mutation_frequency.R
\name{filter_by_mutation_frequency}
\alias{filter_by_mutation_frequency}
\title{Filter a bipartite graph's genes by number of mutations}
\usage{
filter_by_mutation_frequency(bgr, min_times_mut, sample_is = TRUE)
}
\arguments{
\item{bgr}{a bipartite edge graph that has the node attributes \code{type}
and \code{name}; these attributes are not checked for explicitly, but will
throw an error if not present}

\item{min_times_mut}{minimum number of mutations per gene}

\item{sample_is}{either \code{TRUE} or \code{FALSE} dictating which boolean
value corresponds to the samples in the "type" attribute of \code{bgr}}
}
\value{
a vector of (unique) gene names
}
\description{
Returns a vector of genes that passes the minimum number of
  mutations.
}
\examples{
library(wext)
suppressPackageStartupMessages(library(tidygraph))
bgr <- make_sample_gene_bipartite(
    simple_dataset$sample_name,
    simple_dataset$mutated_gene
)
filter_by_mutation_frequency(bgr, 2)
filter_by_mutation_frequency(bgr, 3)
filter_by_mutation_frequency(bgr, 4, sample_is = FALSE)

}
