% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_node_index}
\alias{get_node_index}
\title{Get the indices of nodes after evaluating 'expr'}
\usage{
get_node_index(gr, expr)
}
\arguments{
\item{gr}{tidygraph graph object}

\item{expr}{an expression to evaluate in \code{tidygraph::filter(expr)}}
}
\value{
a vector of the indices of the nodes remaining after the filter
}
\description{
Evalute the expression in \code{tidygraph::filter(expr)} and
  return the indices of the remaining nodes
}
\examples{
suppressPackageStartupMessages(library(magrittr))
suppressPackageStartupMessages(library(tidygraph))
tidygraph::create_ring(5) \%>\%
    mutate(name = LETTERS[1:5]) \%>\%
    get_node_index(name != "B")

}
