% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_gene_bipartite.R
\name{make_sample_gene_bipartite}
\alias{make_sample_gene_bipartite}
\title{Make the sample-gene bipartite graph}
\usage{
make_sample_gene_bipartite(s, g)
}
\arguments{
\item{s}{vector of samples}

\item{g}{vector of mutated genes}
}
\value{
a tidygraph graph object with node attributes \code{name} and
  \code{type}. Attribute \code{type} holds boolean values to indicate which
  of the two groups the node belongs to - \code{TRUE} is for samples.
}
\description{
Create a bipartite graph connecting cancer samples and their
  mutations.
}
\examples{
make_sample_gene_bipartite(
    simple_dataset$sample_name,
    simple_dataset$mutated_gene
)

}
