% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tibble.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{Convert between a mutation matrix and tibble}
\usage{
matrix_to_tibble(dat, columns_name = "sample_id", rows_name = "gene")
}
\arguments{
\item{dat}{a matrix object with column and row names}

\item{columns_name}{title for the data in the column names (default "sample_id")}

\item{rows_name}{title for the data in the row names (default "gene")}
}
\description{
While tibbles are great for interactive data science, matrices
  are more practical for the algebra used by WExT. Therefore, these two
  functions \code{mutationmatrix_to_tibble()} and
  \code{tibble_to_mutationmatrix()}, make switching between the two formats
  much easier.
}
\examples{
set.seed(0)
mat <- matrix(sample(c(0,1), 12, replace = TRUE), nrow = 3)
colnames(mat) <- LETTERS[1:4]
rownames(mat) <- letters[1:3]
matrix_to_tibble(mat)

}
