% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_edge_nodes.R
\name{random_edge_nodes}
\alias{random_edge_nodes}
\alias{random_edge_nodes2}
\title{Return the nodes for a randomly selected edge}
\usage{
random_edge_nodes(gr, ignore_nodes_connected_to = c())

random_edge_nodes2(gr, max_try = 100)
}
\arguments{
\item{gr}{tidygraph graph object; must have the node attribute \code{.idx}
with the node indices}

\item{ignore_nodes_connected_to}{a vector of indices for the nodes whose
edges to ignore}
}
\value{
a vector of two indices for the nodes of the edge selected; if no
  edge was able to be selected, then \code{integer(0)} is returned
}
\description{
Return the nodes of an edge in the graph \code{gr} selected at
  random. Edges from other nodes can be ignored by passing their index(es) to
  \code{ignore_nodes_connected_to}. This is used in swapping edges by passing
  the nodes for the first edge to be ignored in the selection of the second
  edge.
}
\examples{
set.seed(0)
gr <- tidygraph::create_ring(5)
gr <- tidygraph::mutate(gr, .idx = 1:dplyr::n())
random_edge_nodes(gr)

}
