% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_test.R
\name{rc_test}
\alias{rc_test}
\title{The Row-Column-Exclusivity or Co-mutation test}
\usage{
rc_test(dat, sample_col, mutgene_col, k = 2,
  which_test = c("exclusivity", "comutation"), seed_genes = c(),
  N_perms = 10000, min_mutex_events = 2)
}
\arguments{
\item{dat}{tibble with mutation information}

\item{sample_col}{column of samples names (quoted)}

\item{mutgene_col}{column of genes that are mutated (quoted)}

\item{k}{size of gene sets to consider (default is 2)}

\item{which_test}{test for mutual exclusivity (\code{"exclusivity"}) or
co-mutation (\code{"comutation"}); (default is \code{"comutation"})}

\item{seed_genes}{a vector of gene(s) that must be in the gene set to be
tested (optional)}

\item{N_perms}{number of permutation matrices to use (default is 10,000)}

\item{min_mutex_events}{minimum number of real mutual exclusive events
required to consider the gene set (default is 2)}
}
\description{
The Row-Column-Exclusivity or Co-mutation test cannot be
  calculated exactly for most use cases because there is no closed formula
  and a very large number of possible matrices to consider. Thus, the
  probabilities are calculated empirically from a sufficiently large number
  of samples of possible matrices.
}
\examples{
library(wext)
rc_test(simple_dataset, sample_name, mutated_gene, k = 2, N_perms = 5)

}
