% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_edge.R
\name{remove_edge}
\alias{remove_edge}
\title{Remove an edge}
\usage{
remove_edge(gr, a, b)
}
\arguments{
\item{gr}{tidygraph graph object}

\item{a, b}{indices of the two nodes connected by the edge to remove}
}
\value{
a tidygraph graph object without the edge
}
\description{
Remove the edge between \code{a} and \code{b}
}
\examples{
set.seed(0)
plot(remove_edge(tidygraph::create_ring(3), 1, 2))

}
