% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_results_tracker.R
\name{results_tracker}
\alias{results_tracker}
\alias{make_empty_results_tracker}
\alias{exclusivity_results_tracker}
\alias{comutation_results_tracker}
\alias{make_combs_tibble}
\title{Initialize the mutation event tracker tibble}
\usage{
make_empty_results_tracker(bgr, k, which_test, seed_genes, min_times_mut)

exclusivity_results_tracker(bgr, k, seed_genes, min_times_mut)

comutation_results_tracker(bgr, k, seed_genes, min_times_mut)

make_combs_tibble(genes, k, seed_genes)
}
\arguments{
\item{bgr}{a bipartite edge graph that has the node attributes \code{type}
and \code{name}; these attributes are not checked for explicitly, but will
throw an error if not present}

\item{k}{size of gene sets to consider (default is 2)}

\item{which_test}{test for mutual exclusivity (\code{"exclusivity"}) or
co-mutation (\code{"comutation"}); (default is \code{"comutation"})}

\item{seed_genes}{a vector of gene(s) that must be in the gene set to be
tested (optional)}

\item{min_times_mut}{minimum number of times a gene must be mutated in all
samples to be considered for the gene sets (default is 5)}

\item{genes}{a vector of gene(s) to make combinations of size \code{k} from}
}
\value{
a tibble with two columns: \code{gene_sets} contains list objects of
  \code{k} genes, \code{t_BM_gr} is a column of zeros that will eventually
  hold the number of events (either mutual exclusivty or co-mutation) that
  occur between the genes in the gene sets in the permuted bipartite graphs
}
\description{
A tibble that contains the gene sets to consider mutual
  exclusivity or co-mutation between. This uses a bit of logic to reduce the
  number of gene sets to consider by accounting the for test at hand
  (exclusivity or co-mutation), the minimum mutations in a gene to consider
  it in any gene sets, and a seed gene is the user is interested in
  interactions with only one or a few genes.
}
\section{Functions}{
\itemize{
\item \code{exclusivity_results_tracker}: Create the tracker when testing for mutual
exclusivity.

\item \code{comutation_results_tracker}: Create the tracker when testing for co-mutation.

\item \code{make_combs_tibble}: Create the final tibble object with the gene sets
of size \code{k}.
}}

\examples{
library(wext)
bgr <- make_sample_gene_bipartite(
    simple_dataset$sample_name,
    simple_dataset$mutated_gene
)
make_empty_results_tracker(bgr, 2, "exclusivity", c(), 2)

}
