% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_dataset}
\alias{simple_dataset}
\title{Simple example data for WExT}
\format{a tibble (36 x 2)
\describe{
    \item{sample_name}{name of the sample}
    \item{mutated_gene}{mutated gene in the sample}
}}
\source{
\url{https://github.com/raphael-group/wext/tree/master/examples/simple}
}
\usage{
simple_dataset
}
\description{
A small tibble with two columns to indicate the mutations in each sample
}
\keyword{datasets}
