% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_an_edgeC}
\alias{swap_an_edgeC}
\title{Swap a single edge (method 2 in C++)}
\usage{
swap_an_edgeC(n1, n2, N, max_try, quiet)
}
\arguments{
\item{n1, n2}{the graph built as a set of two vectors with the nodes for each edge}

\item{N}{the number of edges to swap (one at a time)}

\item{max_try}{Number of times to try to find two edges to swap. No message
is relayed for a single unsuccessful edge swap; instead a message at the
end prints the number of successful edge swaps in total.}

\item{quiet}{boolean for if you want the number of successful edge swaps
printed}
}
\value{
\code{n1} with nodes swapped
}
\description{
Swap an edge of a bipartite graph while maintaining the
  bipartite nature. This function is not perfectly named as it runs more
  than a single edge swap (whereas the R implemented functions only run a
  single swap).
}
